+++ lib/features/ads/rewarded_interstitial/rewarded_interstitial_helper.dart
@@ -48,7 +48,7 @@ class RewardedInterstitialHelper {
     return completer.future;
   }
 
-  static Future<RewardResult> show({required int rewardAmount, String rewardType = 'coins'}) async {
+  static Future<RewardResult> show({required double rewardAmount, String rewardType = 'coins'}) async {
     if (_rewardedInterstitialAd == null) return RewardResult.failed();
     final completer = Completer<RewardResult>();
     bool earnedReward = false;
